requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
crontab==0.23.0
filelock==3.7.1
huggingface-hub==0.7.0
idna==3.3
joblib==1.1.0
lxml==4.9.0
numpy==1.22.4
oauthlib==3.2.0
packaging==21.3
protobuf==3.20.0
pyparsing==3.0.9
python-dotenv==0.20.0
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
sacremoses==0.0.53
schedule==1.1.0
sentencepiece==0.1.96
six==1.16.0
soupsieve==2.3.2.post1
tokenizers==0.12.1
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cpu
tqdm==4.64.0
transformers==4.20.0
tweepy==4.10.0
typing_extensions==4.2.0
urllib3==1.26.9
